+++ src/main/java/com/h3/reservation/slack/dto/response/factory/modalupdate/ChangeUpdateModalResponseFactory.java
@@ -1,4 +1,4 @@
-package com.h3.reservation.slack.dto.response.factory;
+package com.h3.reservation.slack.dto.response.factory.modalupdate;
 
 import com.h3.reservation.slack.dto.response.ModalUpdateResponse;
 import com.h3.reservation.slack.fragment.block.SectionBlock;
@@ -12,16 +12,16 @@ import java.util.Collections;
  * @version 1.0
  * @date 2019-12-11
  */
-public class ReserveModalUpdateResponseFactory {
+public class ChangeUpdateModalResponseFactory {
     public static ModalUpdateResponse of() {
         return new ModalUpdateResponse(
             new ModalView(
-                "reservation_result",
-                new PlainText("예약하기"),
+                "change_result",
+                new PlainText("변경/취소하기"),
                 new PlainText("확인"),
                 Collections.singletonList(
                     new SectionBlock(
-                        new PlainText("예약하기는 아직 기능이 없습니다. 커밍쑨~")
+                        new PlainText("변경/취소하기는 아직 기능이 없습니다. 커밍쑨~")
                     )
                 )
             )
