+++ src/main/java/com/h3/reservation/slack/dto/response/factory/modalupdate/RetrieveModalUpdateResponseFactory.java
@@ -1,4 +1,4 @@
-package com.h3.reservation.slack.dto.response.factory;
+package com.h3.reservation.slack.dto.response.factory.modalupdate;
 
 import com.h3.reservation.slack.dto.response.ModalUpdateResponse;
 import com.h3.reservation.slack.fragment.block.Block;
