+++ src/main/java/com/h3/reservation/slack/service/SlackService.java
@@ -9,10 +9,10 @@ import com.h3.reservation.slack.dto.request.viewsubmission.ChangeRequest;
 import com.h3.reservation.slack.dto.request.viewsubmission.ReserveRequest;
 import com.h3.reservation.slack.dto.request.viewsubmission.RetrieveRequest;
 import com.h3.reservation.slack.dto.response.ModalUpdateResponse;
-import com.h3.reservation.slack.dto.response.factory.ChangeUpdateModalResponseFactory;
+import com.h3.reservation.slack.dto.response.factory.modalupdate.ChangeUpdateModalResponseFactory;
 import com.h3.reservation.slack.dto.response.factory.InitResponseFactory;
-import com.h3.reservation.slack.dto.response.factory.ReserveModalUpdateResponseFactory;
-import com.h3.reservation.slack.dto.response.factory.RetrieveModalUpdateResponseFactory;
+import com.h3.reservation.slack.dto.response.factory.modalupdate.ReserveModalUpdateResponseFactory;
+import com.h3.reservation.slack.dto.response.factory.modalupdate.RetrieveModalUpdateResponseFactory;
 import com.h3.reservation.slackcalendar.domain.DateTime;
 import com.h3.reservation.slackcalendar.domain.Reservations;
 import com.h3.reservation.slackcalendar.service.SlackCalendarService;
